+++ modules/test/basic.test.js
@@ -14,7 +14,12 @@ test('[Metatest] Schema without errors & warnings', () => {
       $type: 'array',
       $required: false,
       $rules: [ip => ip[0] === '192'], //? custom rules
-      items: { $type: 'union', types: ['string', '?number'], condition: 'oneof', $required: false },
+      items: {
+        $type: 'union',
+        types: ['string', 'number', null],
+        condition: 'oneof',
+        $required: false,
+      },
     },
     type: ['elite', 'member', 'guest'], //? enum
     address: 'string',
@@ -32,7 +37,7 @@ test('[Metatest] Schema without errors & warnings', () => {
       ip: ['192', 168, '1', null],
       type: 'elite',
       mask: ['255', '255', '255', '0'],
-      name: new Set(['Alexander', null]),
+      name: new Set(['Alexander', undefined]),
       options: { notifications: true, lvls: [2, '["admin", "user"]'] },
       associations: { userId: 1, recordId: 1, dbId: 1 },
       address: 'Pushkin street',