+++ modules/test/index.js
@@ -18,26 +18,16 @@ module.exports = (schema, options) => {
     const tests = rules.filter(test => typeof test === 'string' || typeof test === 'function');
     typeof this.test === 'function' && tests.unshift(this.test);
     this.test = (sample, path = 'root', isPartial = false) => {
-      if (sample === undefined || sample === null) {
-        if (!this.$required) return [];
-        return [new Error({ cause: REQUIRE_SAMPLE, path, plan, sample })];
-      }
-
+      const err = (def, cause = def) => new Error({ cause, path, plan, sample });
+      if (sample === undefined) return this.$required ? [err(REQUIRE_SAMPLE)] : [];
       const errors = [];
       for (let i = 0; i < tests.length; i++) {
-        const err = (def, cause = def) => new Error({ cause, path, plan, sample });
         let [rule, name] = [tests[i], i - 1 < 0 ? 'Prototype test' : 'Rule №' + i];
         if (typeof rule === 'string') [rule, name] = [schemaRules.get(rule), rule];
-        if (rule) {
-          const result = rule(sample, path, isPartial);
-          errors.push(...unifyResult(result, err.bind(null, DID_NOT_PASSED + name)));
-          continue;
-        }
-        errors.push(err(RULE_NOT_EXIST + name));
+        const result = rule ? rule(sample, path, isPartial) : RULE_NOT_EXIST + name;
+        errors.push(...unifyResult(result, err.bind(null, DID_NOT_PASSED + name)));
       }
-
-      const result = unifyResult(errors);
-      return Object.assign(result, { valid: result.length === 0 });
+      return Object.assign(errors, { valid: errors.length === 0 });
     };
   }
 };